+++ lib/Client.php
@@ -258,6 +258,11 @@ class Client extends OpenIDConnectClient {
 		return $this->getAutoProvisionConfig()['picture-claim'] ?? null;
 	}
 
+	public function getGroupsClaim(): ?string
+	{
+		return $this->getAutoProvisionConfig()['groups-claim'] ?? null;
+	}
+
 	public function getUserEmail($userInfo): ?string {
 		$email = $this->mode() === 'email' ? $userInfo->{$this->getIdentityClaim()} : null;
 		$emailClaim = $this->getEmailClaim();
@@ -279,6 +284,14 @@ class Client extends OpenIDConnectClient {
 		$pictureClaim = $this->getPictureClaim();
 		if ($pictureClaim) {
 			return $userInfo->$pictureClaim;
+		}
+		return null;
+	}
+
+	public function getUserGroupIds($userInfo): ?string {
+		$groupsClaim = $this->getGroupsClaim();
+		if ($groupsClaim) {
+			return $userInfo->$groupsClaim;
 		}
 		return null;
 	}